// ANGULAR/ANGULAR/src/app/Shared/Model/Master/classification-pagetype-model.ts
export class ClassificationPageTypeModel {
		Id: number;
    PageId: number;
		Code: number | null;
		NameHindi: string;
		NameEnglish: string;
		DisplayOrder: number | null;
		Remarks: string;
		Url: string;
		IsActive: boolean;
		IsDeleted: boolean;
		CreatedDate: Date | string | null;
		CreatedBy: number | null;
		ModifiedDate: Date | string | null;
		ModifiedBy: number | null;
    SampleURl: string;
    AttachmentURl: string;
	}